AMLS-2680: refactoring SubmissionService to put response related things into their own class
